+++ src/jianpu-markdown/errors.ts
@@ -1,0 +1,2 @@
+export const BUG_REPORT_MESSAGE =
+  'This is a bug in Jianpu Markdown parser, we would be grateful if you report it.'