+++ src/music/meta/types.ts
@@ -1,4 +1,4 @@
-import { HtmlNode } from '../../jianpu-markdown/ast'
+import { MarkdownHtmlPropertyValue } from '../../jianpu-markdown/ast'
 import type { Note } from '../music'
 
 export enum Accidental {
@@ -21,5 +21,5 @@ export interface ITimeSignature {
 export interface IMusicMeta {
   key?: IKeySignature
   tempo?: ITimeSignature
-  composer?: string | HtmlNode | HtmlNode[]
+  composer?: MarkdownHtmlPropertyValue
 }